添加SKU管理

+++ model/src/main/java/com/atguigu/gmall/product/entity/SkuInfo.java
@@ -31,7 +31,7 @@ public class SkuInfo implements Serializable {
     /**
      * 价格
      */
-    private Integer price;
+    private BigDecimal price;
     /**
      * sku名称
      */
